+++ src/main/java/stevekung/mods/moreplanets/utils/client/particle/ParticleCustomRain.java
@@ -72,7 +72,7 @@ public class ParticleCustomRain extends Particle
 
             if (state.getBlock() instanceof BlockLiquid)
             {
-                d0 = 1.0F - BlockLiquid.getLiquidHeightPercent(state.getValue(BlockLiquid.LEVEL).intValue());
+                d0 = 1.0F - BlockLiquid.getLiquidHeightPercent(state.getValue(BlockLiquid.LEVEL));
             }
             else
             {
